Explicitly use the v0 server API for all requests
The newest versions of the Chef Server present new behavior for the
user endpoints. Code to handle this new behavior has been added to
core chef. Rather than replicate all that code here, we can use
Chef::ServerAPI to use the v0 API.

The goal is to move all of this plugin into core Chef before v0 is
deprecated.