Remove bugged --no-disassociate-user option for user delete.
Due to a bug, passing --no-disassociate-user had no effect.
This is due to a Chef::Knife convention that switches that start with --no automatically default to a boolean type with the value of false.
Therefore, `unless config[:no_disassociate_user]` would evaluate to `unless nil` if --no-disassociate-user was not passed and `unless false` if passed,
meaning the dissociation code is called regardless of whether or not the user passed the --no-disassociate-user flag.

Intead of fixing the bug, I have removed the never working --no-disassociate-user option.
I chose to remove it because we now have a relevant error message in the admins failure case,
and a --force option for org user remove that can be used to fix any issues.
Furthermore, deleting a user without first dissociatingd can leave an organization in a bad state if that user is the last admin.
Better to force users to properly dissociate from all orgs before deletion (and this is what the code was doing before I removed this flag due to the bug).